pyproject.toml
[project]
name = "fasthtml-auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "monsterui>=1.0.25",
    "python-fasthtml>=0.12.24",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
ignore = ["F403", "F405"]
